test(core): split build options batch e tests
